Add Unit Tests for Infrastructure layer

+++ src/AkhshamBazari.Presentation/Dtos/ProductDto.cs
@@ -3,5 +3,5 @@ namespace AkhshamBazari.Presentation.Dtos;
 public class ProductDto
 {
     public string? Name { get; set; }
-    public decimal? Price { get; set; }
+    public double? Price { get; set; }
 }
